Use Travis for CI
test_plugin_pull_request job on our Jenkins is not yet
in foreman-infra, which is kind of a dead end for me.
I need to remove mysql from test_plugin_matrix because
katello does not support it and there are crashes during
migrations (indexes too long). It can be easily
done with dynamic axis plugin for Jenkins
once the job is in yml file. I decided to run
the tests on Travis to move forward,
which will also allow me greater controll over the process.

Additional complication is possibly broken package
for qpidd on debian. I am not sure what is going on,
but qpid_messaging gem just won't install due to some
missing file in native extensions. So I run the tests
in CentOS 7 container, which is obviously not ideal.
On the other hand, the tests are running.